requirements.txt
Pillow>=10.2.0
terminaltables~=3.1.10
torchsummary~=1.5.1
matplotlib~=3.7.3
torch~=2.0.1
tqdm~=4.65.2
torchvision
numpy~=1.25.2
imgaug~=0.4.0
tensorboard~=2.14.0
opencv-python~=4.8.0.76
thop~=0.1.1.post2209072238
pandas~=2.1.0
pytest~=7.4.3
scikit-image~=0.21.0
scipy~=1.11.2
profilehooks~=1.12.0
scikit-learn~=1.3.1
filterpy~=1.4.5
torcheval~=0.0.7
seaborn
plotly
kaleido
scikit-learn
iterative-stratification
pyyaml